TRex: Move StartServer to client.go
Currently, there is a dedicated file to start the
TRex service.

Consolidate all the TRex functionality under `trex.Client`.
